+++ agent-allure-testng/src/test/java/com/epam/reportportal/testng/AllureIdAnnotationTest.java
@@ -53,6 +53,7 @@ public class AllureIdAnnotationTest {
 	@BeforeEach
 	public void initMocks() {
 		mockLaunch(client, namedUuid("launchUuid"), suitedUuid, testClassUuid, stepUuid);
+		mockLogging(client);
 		ReportPortal reportPortal = ReportPortal.create(client, standardParameters());
 		TestNgListener.REPORT_PORTAL_THREAD_LOCAL.set(reportPortal);
 	}