+++ agent-allure-testng/src/test/java/com/epam/reportportal/testng/MutedAnnotationTest.java
@@ -48,6 +48,7 @@ public class MutedAnnotationTest {
 	@BeforeEach
 	public void initMocks() {
 		mockLaunch(client, namedUuid("launchUuid"), suitedUuid, testClassUuid, stepUuids);
+		mockLogging(client);
 		ReportPortal reportPortal = ReportPortal.create(client, standardParameters());
 		TestNgListener.REPORT_PORTAL_THREAD_LOCAL.set(reportPortal);
 	}